Handle Smart Tags:
Smart Tags are a docx feature that allow you to label text and take
special actions with it (e.g. you can Smart Tag a name so that you
can add it to your outlook contacts from word).

We don't need this info, but the Smart Tagged text is wrapped in an
extra xml element: w:smartTag and is preventing the child run from
being processed. Now the runs inside of the w:smartTag are yielded.