feat: add support for reset numbering
Reset happens when there are two paragraphs on the same level,
with the same style, but have different numIds.
When we encounter reset, check if that paragraph has set numbering
value, and add it to the count.

Optimize getting siblings with generator
Move retrieving ilvl and numId to a function.