fix: `ImagePart` alt text setter not working
Since `image_parts` inside the paragraph are loaded from xml
on each access, setting alt text for `image_part` did not work.
In order to fix it, `docPr` is shared from `InlineShape`
so both instances have the same properties and that way
`descr` (alt text) is same for both instances.