fix: ignore noBreakHyphen when part of `w:fldChar`
This is a partial fix for issue when `w:noBreakHyphen` is shown
when part of complex field group. It is not a full solution since
it checks if `w:noBreakHyphen` is in the same run as `w:instrText`.
Even though `w:instrText` is used in `w:fldChar` it can be used outside
of it as well. In that situations its content is ignored, so it's not
safe to ignore `w:noBreakHyphen` in those situations.

We should in future create a better solution that will actually
properly display conetnets of complex field group.

See: https://github.com/openlawlibrary/python-docx/pull/111
for more details.